==> 1_apiDataFetch.ipynb <==
# Fetch Data through API

import requests
import json
import csv
import pandas as pd
from requests.api import head

url = 'https://data.ny.gov/api/views/4a2x-yp8g/rows.json?accessType=DOWNLOAD'
headers = { 
    'Accept': 'application/json', 
    'Content-Type': 'application/json'
}

#API call to get records
response = requests.request("GET", url, headers=headers,data={})
decoded_data = response.text.encode().decode('utf-8-sig')
data = json.loads(decoded_data)

#API result
columns = data['meta']['view']['columns']
rows = data['data']
excluded_columns = 8

# Convert JSON data to CSV format

# Save to a csv file
with open('Output.csv', 'w', newline='', encoding='utf-8') as outputFile:
    fieldnames = [column['name'] for column in columns][excluded_columns:27] # Extract column names from metadata
    writer = csv.writer(outputFile)
    writer.writerow(fieldnames)
    # Write rows to the CSV file
    for row in rows:
        #print(row)
        writer.writerow(row[excluded_columns:27])   

# Replace null values with NA

csv_data = pd.read_csv('Output.csv')
csv_data.fillna('NA', inplace=True)
csv_data.to_csv('CleanedOutput.csv', index=False)
