==> 2_DB_MySQL.ipynb <==
%pip install mysql-connector-python
%pip install pymysql
%pip install pandas

# Connecting DB in the Debian server through MySQL

import mysql.connector as msql
from mysql.connector import Error
import csv

try:
    conn = msql.connect(
       host="localhost", 
       database="x22105522DB", 
       user="root", 
       password="Karthika@31"
    )
    if conn.is_connected():
      cursor = conn.cursor()
      cursor.execute("select database();")
      record = cursor.fetchone()

      print("You're connected to database: ", record)
      cursor.execute('DROP TABLE IF EXISTS EnergyEfficiencyProject;')

      print('Creating table....')
      cursor.execute("CREATE TABLE EnergyEfficiencyProject (Reporting_Period CHAR(30) NOT NULL, Project_ID CHAR(30) NOT NULL, Project_County CHAR(30) NOT NULL, Project_City CHAR(30) NOT NULL, Project_ZIP CHAR(30) NOT NULL, Gas_Utility CHAR(30) NOT NULL, Electric_Utility CHAR(30) NOT NULL, Project_Completion_Date CHAR(30) NOT NULL, Total_Project_Cost CHAR(30) NOT NULL, Heating_FuelType CHAR(30) NOT NULL, Year_Home_Built CHAR(250) NOT NULL, Size_Of_Home CHAR(30) NOT NULL, Number_Of_Units CHAR(30) NOT NULL, Job_Type CHAR(30) NOT NULL, Type_Of_Dwelling CHAR(30) NOT NULL, Measure_Type CHAR(30) NOT NULL, Annual_Savings_kWh CHAR(30) NOT NULL, Annual_Savings_MMBtu CHAR(30) NOT NULL, Energy_Savings CHAR(30) NOT NULL)")
      print("EnergyEfficiencyProject table is created....")

except Error as e:
    print("Error while connecting to MySQL", e)

# INSERT DATA INTO THE TABLE EnergyEfficiencyProject

# Open the CSV file and read its contents
with open('CleanedEnergyEfficiency.csv', 'r') as f:
    reader = csv.reader(f)
    next(reader) # skip the header row
    for row in reader:
        
        #Insert each row into the database
        cursor = conn.cursor()
        cursor.execute("INSERT INTO EnergyEfficiencyProject (Reporting_Period, Project_ID, Project_County, Project_City, Project_ZIP, Gas_Utility, Electric_Utility, Project_Completion_Date, Total_Project_Cost, Heating_FuelType, Year_Home_Built, Size_Of_Home, Number_Of_Units, Job_Type, Type_Of_Dwelling, Measure_Type, Annual_Savings_kWh, Annual_Savings_MMBtu, Energy_Savings) VALUES (%s, %s, %s,%s, %s, %s,%s, %s, %s,%s, %s, %s,%s, %s, %s,%s, %s, %s,%s);", (row[0], row[1], row[2],row[3],row[4],row[5],row[6],row[7],row[8],row[9],row[10],row[11],row[12],row[13],row[14],row[15],row[16],row[17],row[18]))
        conn.commit()
        print("Table inserted...")

# Close the database connection
conn.close()
