==> 2_X22114386DataBase.ipynb <==
# The code is related to Data base connection
01. Connection string to create a connection pipe line to Database
02. Creation of Database Schema
03. Dump the entire competation.csv file to the table ""
04. Nomalization of the data
05. Insertion of data with respect to tables such as "CountryTable","StateTable","CityTable","PollutionDataTable"

01. Function to Connect the data base where the connection method is TCP/IP over SSH 


%pip install pymysql
%pip install pandas



import csv


def getCsvRowChunks():
    chunkSize = 1000
    csvRowChunks = []

    with open('C:/Users/Manu/Downloads/StudiesNCI/MichealBrad_DB/projects/competitions.csv', 'r', encoding='utf-8') as csvfile:
        # Using DictReader becuase comma in some of the CSV column, normal read will split values based on comma
        csvreader = csv.DictReader(csvfile)
        output = list(csvreader)
        csvRowList = []
        for row in output:
            csvColumnList = []
            for key,val in row.items():
                if(key == "id"):
                    val = int(val)
                csvColumnList.append(val)
            csvRowList.append(csvColumnList)
            
            # chunking csv row lists
            if(len(csvRowList) >= chunkSize):
                csvRowChunks.append(csvRowList)
                csvRowList = []
            
        # Left over
        csvRowChunks.append(csvRowList)
        # print(csvRowChunks)
    return csvRowChunks


import pymysql
import paramiko
import pandas as pd
#import SSHClient
from sshtunnel import SSHTunnelForwarder
from mysql.connector import Error
from datetime import datetime
#from os.path import expanduser



#home = expanduser('~')
ssh_key_path = "C:/Users/Manu/Downloads/StudiesNCI/MichealBrad_DB/projects/DataBase_And_Analytics_Project/new.pem"
mypkey = paramiko.RSAKey.from_private_key_file(ssh_key_path)
#mypkey = paramiko.RSAKey.from_private_key_file("C:/Users/Manu/Downloads/StudiesNCI/MichealBrad_DB/projects/DataBase_And_Analytics_Project/new.pem")
# if you want to use ssh password use - ssh_password='your ssh password', bellow

sql_hostname = '127.0.0.1'
sql_username = 'root'
sql_password = 'root'
sql_main_database = 'x22114386project'
sql_port = 3306
ssh_host = '87.44.4.70'
ssh_user = 'debian'
ssh_port = 22
sql_ip = '0.0.0.0:0'

try:
      tunnel = SSHTunnelForwarder((ssh_host, 22), ssh_pkey=mypkey, ssh_username=ssh_user,
      remote_bind_address=(sql_hostname, 3306)) 
      tunnel.start()
      conn = pymysql.connect(host='127.0.0.1', user='root', passwd='root', database='x22114386project',port=tunnel.local_bind_port)
      cursor=conn.cursor()
     
      print('Dropping table')
      cursor.execute('DROP TABLE IF EXISTS PollutionData;')

      print('Creating table')
      cursor.execute("""CREATE TABLE 
                     PollutionData(id bigint,
                     country varchar(255),
                     state varchar(255),
                     city varchar(255),
                     station varchar(255),
                     last_update varchar(255),
                     pollutant_Gas varchar(255),
                     pollutant_min varchar(255),
                     pollutant_max varchar(255),
                     pollutant_avg varchar(255),
                     pollutant_unit varchar(255));
                     """)


      sql = """INSERT INTO PollutionData
                  (id,
                  country,
                  state,
                  city,
                  station,
                  last_update,
                  pollutant_Gas,
                  pollutant_min,
                  pollutant_max,
                  pollutant_avg,
                  pollutant_unit)
                  VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
            """

      csvRowChunks =  getCsvRowChunks()
      #Insert number of rows = chunk size as a batch
      for csvRowList in csvRowChunks:
         cursor.executemany(sql, csvRowList)
      print('Inserted records')

      conn.commit()
      cursor.close()
      conn.close()

except Error as e:
    print("Error while connecting to MySQL", e)
