==> 3_DataCleaning.ipynb <==
# Data Cleaning

#Check total number of records
import pandas as pd
import numpy as np

df = pd.read_csv('RawData_EnergyEfficiency.csv')
print('Shape of data: ', df.shape) 

# Viewing top rows of the DataFrame

print(df.head())

#Check number of null values in each column

print('Number of null records in data: \n', df.isnull().sum())

# The column "Year Home Built" can be removed since it has multiple null values

df = pd.read_csv('RawData_EnergyEfficiency.csv')
dataframe = df.drop('Year Home Built', axis=1)

print('Altered Dataframe:')
print(dataframe)

# Replacing empty values with NA

print(dataframe.fillna('NA', inplace=True))

# Generating new csv file with cleaned data
dataframe.to_csv('CleanedEnergyEfficiency.csv', index=False)
