==> 6_MergedDatasets.ipynb <==
# Merge all datasets for further visualisation analysis

import pandas as pd

csv_files = ["CleanedTreePlantations.csv", "CleanVehData.csv", "CleanedEnergyEfficiency.csv", "CleanedCompetitions.csv"]

# Combine the CSV files into a single DataFrame
df = pd.concat([pd.read_csv(file) for file in csv_files])

# Write the merged DataFrame to a new CSV file
df.to_csv("MergedData.csv", index=False)


# Define a list of the four CSV file names
file_names = ["CleanedTreePlantations.csv", "CleanVehData.csv", "CleanedEnergyEfficiency.csv", "CleanedCompetitions.csv"]


# Initialize an empty list to store the data frames
dfs = []

# Loop through the file names and load the first 1000 rows of each CSV file into a data frame
for file_name in file_names:
    df = pd.read_csv(file_name, nrows=1000)
    dfs.append(df)


# Concatenate the data frames vertically (along the rows axis)
merged_df = pd.concat(dfs, axis=1)

merged_df.fillna('NA', inplace=True)
# Write the merged dataset to a CSV file
merged_df.to_csv('MergedDataset.csv', index=False)
